# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "thinking calendar API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/dashboard/month": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "看板"
                ],
                "summary": "获取指定月份看板数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "月份，格式YYYY-MM",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/dashboard/summary": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "看板"
                ],
                "summary": "获取看板汇总数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
        "/records": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "date 为空返回当前用户全部记录，传 date 返回单日记录（不存在返回 null）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作记录"
                ],
                "summary": "查询工作记录（单日或全部）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "日期，格式 YYYY-MM-DD",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "同一日期多次调用视为更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作记录"
                ],
                "summary": "创建或更新工作记录",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpsertRecordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.RecordItem"
                        }
                    }
                }
            }
        },
        "/records/range": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "start 和 end 需为 YYYY-MM-DD，且 start \u003c end",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作记录"
                ],
                "summary": "按时间范围查询工作记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.RecordItem"
                            }
                        }
                    }
                }
            }
        },
        "/records/{record_id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作记录"
                ],
                "summary": "删除工作记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录 ID",
                        "name": "record_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v1.RecordItem": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "工作内容",
                    "type": "string",
                    "example": "完成接口定义与联调"
                },
                "date": {
                    "description": "日期，格式 YYYY-MM-DD",
                    "type": "string",
                    "example": "2025-12-11"
                },
                "record_id": {
                    "description": "唯一标识",
                    "type": "string",
                    "example": "rec_123"
                },
                "updatedAt": {
                    "description": "最近更新时间",
                    "type": "string",
                    "example": "2025-12-11T10:00:00Z"
                },
                "version": {
                    "description": "版本计数",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "v1.UpsertRecordReq": {
            "type": "object",
            "required": [
                "content",
                "date"
            ],
            "properties": {
                "content": {
                    "description": "记录内容",
                    "type": "string"
                },
                "date": {
                    "description": "记录日期",
                    "type": "string",
                    "example": "2025-12-11"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}